implemented complete purchases

+++ app/src/main/java/com/example/supermercado/external/appscript/model/AppScriptTokenDto.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class AppScriptTokenDto(
     val token: String,
-    val expires_in: Int,
+    val expires_in: Long,
 )
