+++ app/src/main/java/com/example/supermercado/model/Token.kt
@@ -2,5 +2,5 @@ package com.example.supermercado.model
 
 data class Token(
     val value: String,
-    val expiresIn: Int,
+    val expiresIn: Long,
 )
